Moved read_df and create_df in hdf5.File
