+++ openquake/calculators/base.py
@@ -710,7 +710,7 @@ class HazardCalculator(BaseCalculator):
         if len(self.crmodel):
             logging.info('Storing risk model')
             attrs = self.crmodel.get_attrs()
-            self.datastore.create_dframe('crm', self.crmodel.to_dframe(),
+            self.datastore.create_df('crm', self.crmodel.to_dframe(),
                                          'gzip', **attrs)
 
     def _read_risk_data(self):
@@ -1172,7 +1172,7 @@ def create_gmf_data(dstore, prim_imts, sec_imts=(), data=None):
         eff_time = oq.investigation_time * oq.ses_per_logic_tree_path * R
     else:
         eff_time = 0
-    dstore.create_dframe('gmf_data', items, 'gzip')
+    dstore.create_df('gmf_data', items, 'gzip')
     dstore.set_attrs('gmf_data', num_events=len(dstore['events']),
                      imts=' '.join(map(str, prim_imts)),
                      effective_time=eff_time)
