+++ openquake/calculators/event_based_risk.py
@@ -283,14 +283,14 @@ class EventBasedRiskCalculator(event_based.EventBasedCalculator):
         if 'risk' in oq.calculation_mode:
             descr = [('event_id', U32), ('agg_id', U32), ('loss_id', U8),
                      ('loss', F32), ('variance', F32)]
-            self.datastore.create_dframe(
+            self.datastore.create_df(
                 'agg_loss_table', descr,
                 K=len(self.aggkey), L=len(oq.loss_names))
         else:  # damage
             dmgs = ' '.join(self.crmodel.damage_states[:1])
             descr = ([('event_id', U32), ('agg_id', U32), ('loss_id', U8)] +
                      [(dc, F32) for dc in self.crmodel.get_dmg_csq()])
-            self.datastore.create_dframe(
+            self.datastore.create_df(
                 'agg_loss_table', descr,
                 K=len(self.aggkey), L=len(oq.loss_names), limit_states=dmgs)
         ws = self.datastore['weights']
